requirements.txt
black==22.6.0
boto3==1.24.76
django==4.1
django-environ==0.9.0
django-storages==1.13.1
flake8==5.0.4
icecream
psycopg2-binary==2.9.3
pillow==9.2.0
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-selenium==3.0.0
whitenoise==6.2.0
